pyproject.toml
[project]
name = "llm-benchmark-datasets"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.48.2",
    "fastparquet>=2024.11.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "weave>=0.52.20",
]
